@@ -15,13 +15,14 @@
   ],
   "type": "yii2-extension",
   "require": {
-    "php": ">=5.4.0",
-    "yiisoft/yii2": "2.0.*",
-    "symfony/process": "2.6.* || 3.* || 4.*",
-    "dragonmantank/cron-expression": "1.*"
+    "php": ">=8.4.0",
+    "yiisoft/yii2": "^2.0.53",
+    "symfony/process": "2.6.* || 3.* || 4.* || 5.* || 6.* || 7.*",
+    "dragonmantank/cron-expression": "1.*",
+    "guzzlehttp/guzzle": "8.0.x-dev"
   },
   "require-dev": {
-    "phpunit/phpunit": "4.8.36"
+    "phpunit/phpunit": "12.3.*"
   },
   "suggest": {
     "guzzlehttp/guzzle": "Required to use the thenPing method on schedules (~5.0)."
