@@ -1,9 +1,11 @@
 <?php
 
 namespace omnilight\scheduling;
+
+use Override;
 use Yii;
 use yii\base\Application;
-use yii\base\InvalidParamException;
+use yii\base\InvalidArgumentException;
 use yii\mutex\Mutex;
 
 /**
@@ -16,13 +18,14 @@ class CallbackEvent extends Event
      *
      * @var string
      */
-    protected $callback;
+    protected string $callback;
+
     /**
      * The parameters to pass to the method.
      *
      * @var array
      */
-    protected $parameters;
+    protected array $parameters;
 
     /**
      * Create a new event instance.
@@ -31,21 +34,16 @@ class CallbackEvent extends Event
      * @param string $callback
      * @param array $parameters
      * @param array $config
-     * @throws InvalidParamException
+     * @throws InvalidArgumentException
      */
-    public function __construct(Mutex $mutex, $callback, array $parameters = [], $config = [])
+    public function __construct(Mutex $mutex, string $callback, array $parameters = [], array $config = [])
     {
         $this->callback = $callback;
         $this->parameters = $parameters;
-        $this->_mutex = $mutex;
+        parent::__construct($mutex, config: $config);
 
-        if (!empty($config)) {
-            Yii::configure($this, $config);
-        }
-
-        if ( ! is_string($this->callback) && ! is_callable($this->callback))
-        {
-            throw new InvalidParamException(
+        if (!is_string($this->callback) && !is_callable($this->callback)) {
+            throw new InvalidArgumentException(
                 "Invalid scheduled callback event. Must be string or callable."
             );
         }
@@ -57,7 +55,7 @@ class CallbackEvent extends Event
      * @param Application $app
      * @return mixed
      */
-    public function run(Application $app)
+    public function run(Application $app): mixed
     {
         $this->trigger(self::EVENT_BEFORE_RUN);
         $response = call_user_func_array($this->callback, array_merge($this->parameters, [$app]));
@@ -70,12 +68,12 @@ class CallbackEvent extends Event
      * Do not allow the event to overlap each other.
      *
      * @return $this
-     * @throws InvalidParamException
+     * @throws InvalidArgumentException
      */
-    public function withoutOverlapping()
+    public function withoutOverlapping(): static
     {
         if (empty($this->_description)) {
-            throw new InvalidParamException(
+            throw new InvalidArgumentException(
                 "A scheduled event name is required to prevent overlapping. Use the 'description' method before 'withoutOverlapping'."
             );
         }
@@ -88,7 +86,7 @@ class CallbackEvent extends Event
      *
      * @return string
      */
-    protected function mutexName()
+    protected function mutexName(): string
     {
         return 'framework/schedule-' . sha1($this->_description);
     }
@@ -98,7 +96,7 @@ class CallbackEvent extends Event
      *
      * @return string
      */
-    public function getSummaryForDisplay()
+    public function getSummaryForDisplay(): string
     {
         if (is_string($this->_description)) return $this->_description;
         return is_string($this->callback) ? $this->callback : 'Closure';
