@@ -5,7 +5,9 @@ namespace omnilight\scheduling;
 use Yii;
 use yii\base\Component;
 use yii\base\Application;
+use yii\base\InvalidConfigException;
 use yii\mutex\FileMutex;
+use yii\mutex\Mutex;
 
 
 /**
@@ -14,27 +16,28 @@ use yii\mutex\FileMutex;
 class Schedule extends Component
 {
     /**
-     * All of the events on the schedule.
+     * All the events on the schedule.
      *
      * @var Event[]
      */
-    protected $_events = [];
+    protected array $_events = [];
 
     /**
      * The mutex implementation.
      *
-     * @var \yii\mutex\Mutex
+     * @var Mutex|FileMutex|null
      */
-    protected $_mutex;
+    protected Mutex|FileMutex|null $_mutex;
 
     /**
      * @var string The name of cli script
      */
-    public $cliScriptName = 'yii';
+    public string $cliScriptName = 'yii';
 
     /**
      * Schedule constructor.
      * @param array $config
+     * @throws InvalidConfigException
      */
     public function __construct(array $config = [])
     {
@@ -46,11 +49,11 @@ class Schedule extends Component
     /**
      * Add a new callback event to the schedule.
      *
-     * @param  string  $callback
-     * @param  array   $parameters
-     * @return Event
+     * @param string $callback
+     * @param array $parameters
+     * @return CallbackEvent
      */
-    public function call($callback, array $parameters = array())
+    public function call(string $callback, array $parameters = array()): CallbackEvent
     {
         $this->_events[] = $event = new CallbackEvent($this->_mutex, $callback, $parameters);
         return $event;
@@ -58,10 +61,10 @@ class Schedule extends Component
     /**
      * Add a new cli command event to the schedule.
      *
-     * @param  string  $command
+     * @param string $command
      * @return Event
      */
-    public function command($command)
+    public function command(string $command): Event
     {
         return $this->exec(PHP_BINARY . ' ' . $this->cliScriptName . ' ' . $command);
     }
@@ -69,27 +72,27 @@ class Schedule extends Component
     /**
      * Add a new command event to the schedule.
      *
-     * @param  string  $command
+     * @param string $command
      * @return Event
      */
-    public function exec($command)
+    public function exec(string $command): Event
     {
         $this->_events[] = $event = new Event($this->_mutex, $command);
         return $event;
     }
 
-    public function getEvents()
+    public function getEvents(): array
     {
         return $this->_events;
     }
 
     /**
-     * Get all of the events on the schedule that are due.
+     * Get all the events on the schedule that are due.
      *
-     * @param \yii\base\Application $app
+     * @param Application $app
      * @return Event[]
      */
-    public function dueEvents(Application $app)
+    public function dueEvents(Application $app): array
     {
         return array_filter($this->_events, function(Event $event) use ($app)
         {
