@@ -1,6 +1,9 @@
 <?php
 
 namespace omnilight\scheduling;
+
+use Yii;
+use yii\base\InvalidConfigException;
 use yii\console\Controller;
 use yii\di\Instance;
 
@@ -11,48 +14,53 @@ use yii\di\Instance;
 class ScheduleController extends Controller
 {
     /**
-     * @var Schedule
+     * @var string|Schedule
      */
-    public $schedule = 'schedule';
+    public string|Schedule $schedule = 'schedule';
+
     /**
-     * @var string Schedule file that will be used to run schedule
+     * @var string|null Schedule file that will be used to run schedule
      */
-    public $scheduleFile;
+    public string|null $scheduleFile;
 
     /**
      * @var bool set to true to avoid error output
      */
-    public $omitErrors = false;
+    public bool $omitErrors = false;
 
-    public function options($actionID)
+    public function options($actionID): array
     {
         return array_merge(parent::options($actionID),
             $actionID == 'run' ? ['scheduleFile', 'omitErrors'] : []
         );
     }
 
-
-    public function init()
+    /**
+     * @throws InvalidConfigException
+     */
+    public function init(): void
     {
-        if (\Yii::$app->has($this->schedule)) {
+        if (Yii::$app->has($this->schedule)) {
             $this->schedule = Instance::ensure($this->schedule, Schedule::className());
         } else {
-            $this->schedule = \Yii::createObject(Schedule::className());
+            $this->schedule = Yii::createObject(Schedule::className());
         }
         parent::init();
     }
 
-
-    public function actionRun()
+    /**
+     * @throws InvalidConfigException
+     */
+    public function actionRun(): void
     {
         $this->importScheduleFile();
 
-        $events = $this->schedule->dueEvents(\Yii::$app);
+        $events = $this->schedule->dueEvents(Yii::$app);
 
         foreach ($events as $event) {
             $event->omitErrors($this->omitErrors);
             $this->stdout('Running scheduled command: '.$event->getSummaryForDisplay()."\n");
-            $event->run(\Yii::$app);
+            $event->run(Yii::$app);
         }
 
         if (count($events) === 0)
@@ -61,14 +69,14 @@ class ScheduleController extends Controller
         }
     }
 
-    protected function importScheduleFile()
+    protected function importScheduleFile(): void
     {
         if ($this->scheduleFile === null) {
             return;
         }
 
-        $scheduleFile = \Yii::getAlias($this->scheduleFile);
-        if (file_exists($scheduleFile) == false) {
+        $scheduleFile = Yii::getAlias($this->scheduleFile);
+        if (!file_exists($scheduleFile)) {
             $this->stderr('Can not load schedule file '.$this->scheduleFile."\n");
             return;
         }
