@@ -3,30 +3,33 @@
 namespace omnilight\scheduling\Tests;
 
 use omnilight\scheduling\Event;
+use PHPUnit\Framework\Attributes\DataProvider;
+use PHPUnit\Framework\MockObject\Exception;
+use PHPUnit\Framework\TestCase;
 use yii\mutex\Mutex;
 
-class EventTest extends \PHPUnit_Framework_TestCase
+class EventTest extends TestCase
 {
-    public function buildCommandData()
+    public static function buildCommandData()
     {
         return [
-            [false, 'php -i', '/dev/null', 'php -i > /dev/null'],
-            [false, 'php -i', '/my folder/foo.log', 'php -i > /my folder/foo.log'],
+            [false, 'php -i', '/dev/null', 'php -i > /dev/null &'],
+            [false, 'php -i', '/my folder/foo.log', 'php -i > /my folder/foo.log &'],
             [true, 'php -i', '/dev/null', 'php -i > /dev/null 2>&1 &'],
             [true, 'php -i', '/my folder/foo.log', 'php -i > /my folder/foo.log 2>&1 &'],
         ];
     }
 
     /**
-     * @dataProvider buildCommandData
      * @param bool $omitErrors
      * @param string $command
      * @param string $outputTo
      * @param string $result
+     * @throws Exception
      */
-    public function testBuildCommandSendOutputTo($omitErrors, $command, $outputTo, $result)
+    #[DataProvider('buildCommandData')] public function testBuildCommandSendOutputTo(bool $omitErrors, string $command, string $outputTo, string $result)
     {
-        $event = new Event($this->getMock(Mutex::className()), $command);
+        $event = new Event($this->createMock(Mutex::class), $command);
         $event->omitErrors($omitErrors);
         $event->sendOutputTo($outputTo);
         $this->assertSame($result, $event->buildCommand());
